add RatingBar.jsx
